# swagger.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
  /providers:
    post:
      operationId: ProvidersController_createProvider
      summary: Create a new provider
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProviderDto'
      responses:
        '201':
          description: The create provider request has been successfully processed
      tags: &ref_0
        - provider
      security: &ref_1
        - bearer: []
    get:
      operationId: ProvidersController_readSupplier
      summary: Get providers
      parameters:
        - name: code
          required: false
          in: query
          schema:
            type: string
        - name: status
          required: false
          in: query
          schema:
            type: string
        - name: sortField
          required: false
          in: query
          schema:
            type: string
        - name: sortType
          required: false
          in: query
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - name: offset
          required: false
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Returns an array of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Provider'
      tags: *ref_0
      security: *ref_1
  /providers/{id}:
    put:
      operationId: ProvidersController_updateProvider
      summary: Update a provider
      parameters:
        - name: id
          required: true
          in: path
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                code:
                  type: string
                apiKey:
                  type: string
                url:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
      tags: *ref_0
      security: *ref_1
    delete:
      operationId: ProvidersController_deleteProvider
      summary: Delete a provider
      parameters:
        - name: id
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: Provider deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/'
      tags: *ref_0
      security: *ref_1
  /transactions:
    post:
      operationId: TransactionsController_createTransaction
      summary: Create a new transaction
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionDto'
      responses:
        '201':
          description: The create transaction request has been successfully processed
      tags:
        - transactions
      security:
        - bearer: []
info:
  title: Digital Backend Assignment
  description: The Digital Backend Assignment API description
  version: '1.0'
  contact: {}
tags:
  - name: Digital Backend Assignment
    description: ''
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
    CreateProviderDto:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        url:
          type: string
        username:
          type: string
        password:
          type: string
        apiKey:
          type: string
      required:
        - name
        - code
        - url
    UpdateResult:
      type: object
      properties: {}
    Provider:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        code:
          type: string
        url:
          type: string
        username:
          type: string
        password:
          type: string
        apiKey:
          type: string
        status:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - code
        - url
        - username
        - password
        - apiKey
        - status
        - createdAt
        - updatedAt
    CreateTransactionDto:
      type: object
      properties:
        supplierIds:
          type: array
          items:
            type: number
        departure_post_code:
          type: number
        arrival_post_code:
          type: number
        departure_state_name:
          type: string
        departure_country_code:
          type: string
        arrival_state_name:
          type: string
        arrival_country_code:
          type: string
        item_length:
          type: number
        item_width:
          type: number
        item_height:
          type: number
        item_weight:
          type: number
        addons:
          type: object
      required:
        - supplierIds
        - departure_post_code
        - arrival_post_code
        - departure_state_name
        - departure_country_code
        - arrival_state_name
        - arrival_country_code
        - item_length
        - item_width
        - item_height
        - item_weight
        - addons